enhancement on baby details screen

+++ data/repository/src/main/java/com/aa/repository/resources/LoginResponseResource.kt
@@ -7,7 +7,7 @@ data class LoginResponseResource(
     @SerializedName("access_token")
     val accessToken: String? = null,
     @SerializedName("expires_in")
-    val expiresIn: Int? = null,
+    val expiresIn: Long? = null,
     @SerializedName("token_type")
     val tokenType: String? = null,
     @SerializedName("user")
